requirements.txt
numpy==1.14.5
opencv-python==3.4.1.15
Pillow==5.1.0
scipy==1.1.0
